still more 64-bit fixes for incorrect downcasting

+++ bk2ch08p398cellBackgroundLayering/ch21p701cellBackgroundLayering/RootViewController.m
@@ -60,7 +60,7 @@
 
         
     }
-    cell.textLabel.text = [NSString stringWithFormat:@"Hello there! %d", indexPath.row];
+    cell.textLabel.text = [NSString stringWithFormat:@"Hello there! %ld", (long)indexPath.row];
 
     
     return cell;
